file downloading & pdf generation

+++ controllers/admin.js
@@ -1,5 +1,7 @@
 const Product = require('../models/product');
 
+const fileHelper = require('../util/file')
+
 const { validationResult } = require('express-validator') 
 
 exports.getAddProduct = (req, res, next) => {
@@ -135,6 +137,7 @@ exports.postEditProduct = (req, res, next) => {
 		product.price = updatedPrice;
 		product.description = updatedDesc;
 		if(image){
+			fileHelper.deleteFile(product.imageUrl)
 			product.imageUrl = image.path;
 		}
 		product.save()
@@ -169,7 +172,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
 	const prodId = req.body.productId;
-	Product.deleteOne({_id: prodId, userId: req.user._id})
+	Product.findById(prodId)
+	.then(product => {
+		if(!product) {
+			return next(new Error('Product Not Found'))
+		}
+		fileHelper.deleteFile(product.imageUrl)
+		return Product.deleteOne({_id: prodId, userId: req.user._id})
+	})
 	.then(() => {
 		console.log('DESTROYED PRODUCT');
 		res.redirect('/admin/products');
